test(workbench-cli): Move workbench integration tests to their own project
