+++ packages/@sanity/workbench-cli/test/integration/actions/dev/processLiveness.test.ts
@@ -11,14 +11,14 @@ import {join} from 'node:path'
 
 import {afterEach, beforeEach, describe, expect, test, vi} from 'vitest'
 
-import {__resetStartTimeCacheForTesting} from '../processLiveness.js'
+import {__resetStartTimeCacheForTesting} from '../../../../src/actions/dev/processLiveness.js'
 import {
   acquireWorkbenchLock,
   type DevServerManifest,
   getRegisteredServers,
   readWorkbenchLock,
   registerDevServer,
-} from '../registry.js'
+} from '../../../../src/actions/dev/registry.js'
 
 const mockExecSync = vi.hoisted(() => vi.fn())
