+++ packages/@sanity/workbench-cli/test/integration/actions/dev/workbenchLock.test.ts
@@ -4,8 +4,8 @@ import {join} from 'node:path'
 
 import {afterEach, beforeEach, describe, expect, test, vi} from 'vitest'
 
-import {__resetStartTimeCacheForTesting} from '../processLiveness.js'
-import {acquireWorkbenchLock, readWorkbenchLock} from '../registry.js'
+import {__resetStartTimeCacheForTesting} from '../../../../src/actions/dev/processLiveness.js'
+import {acquireWorkbenchLock, readWorkbenchLock} from '../../../../src/actions/dev/registry.js'
 
 const mockExecSync = vi.hoisted(() => vi.fn())
