+++ vitest.config.ts
@@ -61,7 +61,8 @@ export default defineConfig({
       'packages/@sanity/cli/vitest.config.integration.ts',
       'packages/@sanity/cli-build',
       'packages/@sanity/cli-core',
-      'packages/@sanity/workbench-cli',
+      'packages/@sanity/workbench-cli/vitest.config.ts',
+      'packages/@sanity/workbench-cli/vitest.config.integration.ts',
       'packages/create-sanity',
     ],
     reporters: ['default', ...(IS_AGENT ? ['json'] : [])],
